% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiza.R
\name{visualiza}
\alias{visualiza}
\title{Gráfico de densidade ou histograma}
\usage{
visualiza(df, variavel, tipo = c("densidade", "histograma"))
}
\arguments{
\item{df}{Um objeto do tipo Data.Frame.}

\item{variavel}{Uma coluna do Data.Frame especificado em df.}

\item{tipo}{Uma string sendo "densidade" ou "histograma".}
}
\value{
Um gráfico do `ggplot` de densidade ou histograma de `variavel`
}
\description{
Gráfico de densidade ou histograma
}
\examples{
visualiza(iris, "Sepal.Length", "densidade")
visualiza(iris, "Sepal.Width", "histograma")


}
